feat(fireblocks-signing-manager): allow secret to passed directly instead of requiring a file
initalizes fireblocks by accepting a secret param directly. previously a file path would need to be
given, which would be read in the contructor. this allows for more flexible secret storage

BREAKING CHANGE: `secretPath` is now `secret`. To maintain old behavior use `fs.readFileSync` and
pass the secret to the constructor instead of the path